lib/regurgitate: change scaling algorithm
Replace the "k2" cluster scaling algorithm from the original t-digest
paper with the Folly algorithm. This algorithm is quadratic instead of
logarithmic, so the cluster values are computationally easier to
compute. It also generates less tight bounds near the extremes, so the
resulting digest is easier to map to a traditional histogram.

Additionally, drop a couple of problematic unit tests that were
sensitive to the cluster algorithm due to float resolution issues.

Finally, loosen the AAT check to avoid float resolution issues.

+++ tests/aat/common/matcher/packet_analyzer.py
@@ -88,7 +88,7 @@ class _be_valid_packet_analyzer_flow_digest_result(Matcher):
             for i in range(len(result.centroids) - 1):
                 left = result.centroids[i];
                 right = result.centroids[i + 1]
-                expect(right.mean).to(be_above_or_equal(left.mean))
+                expect(int(right.mean)).to(be_above_or_equal(int(left.mean)))
         return True, ['is valid packet analyzer flow digest result']
